+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # TBD
 
+## Enhancements
+
+- Add method to download device logs from BrowserStack [#344](https://github.com/bugsnag/maze-runner/pull/344)
+
 ## Fixes
 
 - Correct BrowserStack capability to `disableAnimations` [#323](https://github.com/bugsnag/maze-runner/pull/323)